Refactor VOC discovery into dedicated feature module
